fixes/add to pricing page
