+++ src/components/.mdxignore
@@ -81,7 +81,7 @@ Features
 OtherOptions
 Pipelines
 PipelinesList
-PlanComparison
+Plans
 PricingCalculator
 PricingHero
 PricingSlider
