--- src/components/Pricing/PlanComparison/styles/index.scss
@@ -1,7 +1,0 @@
-.pricing-table-collapsed tbody tr:last-of-type {
-    opacity: 0.2;
-}
-
-.pricing-table-expanded tbody tr:last-of-type {
-    opacity: 1;
-}
