+++ src/components/Pricing/Plans/index.tsx
@@ -1,0 +1,370 @@
+import { graphql, useStaticQuery } from 'gatsby'
+import { capitalize } from 'instantsearch.js/es/lib/utils'
+import React, { useState } from 'react'
+import { Check2, Close } from 'components/Icons'
+import Tooltip from 'components/Tooltip'
+import { TrackedCTA } from 'components/CallToAction'
+import usePostHog from 'hooks/usePostHog'
+import Label from 'components/Label'
+import { BillingProductV2Type, BillingV2FeatureType } from 'types'
+
+const Heading = ({ title, subtitle, className = '' }: { title?: string; subtitle?: string; className?: string }) => {
+    return (
+        <div className={className}>
+            <h4 className="m-0 text-base opacity-70">{title}</h4>
+            {subtitle && <p className="m-0 text-sm opacity-70">{subtitle}</p>}
+        </div>
+    )
+}
+
+const Row = ({ children, className = '' }: { children: React.ReactNode; className?: string }) => {
+    return <div className={`grid grid-cols-4 items-center gap-x-4 px-2 py-1.5 rounded ${className}`}>{children}</div>
+}
+
+const Feature = ({ feature }: { feature: BillingV2FeatureType }) => {
+    return feature ? (
+        feature?.limit || feature?.note ? (
+            <p className="m-0 text-base opacity-70">
+                {feature.note || `${feature.limit.toLocaleString()} ${feature.unit}`}
+            </p>
+        ) : (
+            <>
+                <Check2 className="text-green w-5" />
+                <span className="sr-only">Included</span>
+            </>
+        )
+    ) : (
+        <>
+            <Close opacity={1} className="text-red w-4" />
+            <span className="sr-only">Not included</span>
+        </>
+    )
+}
+
+const Title = ({ title, className = '' }: { title: string; className?: string }) => {
+    return <h5 className={`m-0 text-base opacity-70 font-medium ${className}`}>{title}</h5>
+}
+
+const InclusionOnlyRow = ({ plans }) => (
+    <Row className="!py-1">
+        <div className="col-span-2" />
+        {plans.map(({ included_if, plan_key }, index) => (
+            <Title
+                key={`inlclusion-only-${plan_key}-${index}`}
+                title={included_if === 'no_active_subscription' ? 'Free' : 'Paid'}
+                className="font-bold"
+            />
+        ))}
+    </Row>
+)
+
+const PricingTiers = ({ plans, unit, compact = false }) =>
+    plans[plans.length - 1]?.tiers?.map(({ up_to, unit_amount_usd }, index) => {
+        return compact && parseFloat(unit_amount_usd) <= 0 ? null : (
+            <Row className={`!py-1 ${compact ? '!px-0 !gap-x-0' : ''}`} key={`type-${index}`}>
+                <Title
+                    className={`col-span-2 ${compact ? 'text-sm' : ''}`}
+                    title={
+                        index === 0
+                            ? `First ${formatCompactNumber(up_to)} ${unit}s`
+                            : !up_to
+                            ? `${formatCompactNumber(plans[plans.length - 1].tiers[index - 1].up_to)} +`
+                            : `${
+                                  formatCompactNumber(plans[plans.length - 1].tiers[index - 1].up_to).split(' ')[0]
+                              }-${formatCompactNumber(up_to)}`
+                    }
+                />
+                {!compact && <Title className="font-bold" title={plans[0].free_allocation === up_to ? 'Free' : '-'} />}
+                <Title
+                    className={`font-bold ${compact ? 'text-sm col-span-2' : ''}`}
+                    title={
+                        plans[0].free_allocation === up_to
+                            ? 'Free'
+                            : `$${parseFloat(unit_amount_usd).toFixed(
+                                  Math.max(
+                                      ...plans[plans.length - 1].tiers.map((tier) => tier.unit_amount_usd.length)
+                                  ) - 1
+                              )}`
+                    }
+                />
+            </Row>
+        )
+    })
+
+const formatCompactNumber = (number) => {
+    const formatter = Intl.NumberFormat('en', { notation: 'compact', compactDisplay: 'long' })
+    return formatter.format(number)
+}
+
+const AddonTooltipContent = ({ addon }) => {
+    const referencePlan = addon.plans?.[0]
+    const tiers = referencePlan?.tiers
+    const isFirstTierFree = parseFloat(tiers?.[0].unit_amount_usd || '') === 0
+    const [showDiscounts, setShowDiscounts] = useState(false)
+
+    return (
+        <div className="p-2 max-w-sm">
+            <p className="font-bold text-[15px] mb-2">
+                {addon.name} <Label className="ml-2" text="Addon" />
+            </p>
+            <p className="text-sm mb-3">{addon.description}</p>
+            <p className="text-sm opacity-70 mb-3">
+                {isFirstTierFree &&
+                    `First ${formatCompactNumber(tiers?.[0].up_to)} ${referencePlan.unit}s/mo free, then `}
+                <span className="font-bold text-base text-primary dark:text-primary-dark/75">
+                    ${parseFloat((isFirstTierFree ? tiers?.[1]?.unit_amount_usd : tiers?.[0]?.unit_amount_usd) || '')}
+                </span>
+            </p>
+            {showDiscounts ? (
+                <PricingTiers compact unit={addon.unit} plans={addon.plans} />
+            ) : (
+                <button onClick={() => setShowDiscounts(true)} className="text-red dark:text-yellow font-bold">
+                    Show volume discounts
+                </button>
+            )}
+        </div>
+    )
+}
+
+const AddonTooltip = ({ children, addon }: { children: React.ReactNode; addon: BillingProductV2Type }) => {
+    return (
+        <Tooltip placement="right-end" content={() => <AddonTooltipContent addon={addon} />}>
+            <span className="relative">{children}</span>
+        </Tooltip>
+    )
+}
+
+const CTA = () => {
+    const posthog = usePostHog()
+    return (
+        <TrackedCTA
+            event={{
+                name: `clicked Get started - free`,
+                type: 'cloud',
+            }}
+            type="primary"
+            size="md"
+            className="shadow-md !w-auto"
+            to={`https://${
+                posthog?.isFeatureEnabled && posthog?.isFeatureEnabled('direct-to-eu-cloud') ? 'eu' : 'app'
+            }.posthog.com/signup`}
+        >
+            Get started - free
+        </TrackedCTA>
+    )
+}
+
+const allProductsData = graphql`
+    query GetAllProductData {
+        allProductData {
+            nodes {
+                products {
+                    description
+                    docs_url
+                    image_url
+                    inclusion_only
+                    contact_support
+                    addons {
+                        contact_support
+                        description
+                        docs_url
+                        image_url
+                        inclusion_only
+                        name
+                        type
+                        unit
+                        plans {
+                            description
+                            docs_url
+                            image_url
+                            name
+                            plan_key
+                            product_key
+                            unit
+                            features {
+                                description
+                                key
+                                name
+                            }
+                            tiers {
+                                current_amount_usd
+                                current_usage
+                                flat_amount_usd
+                                unit_amount_usd
+                                up_to
+                            }
+                        }
+                    }
+                    name
+                    type
+                    unit
+                    usage_key
+                    plans {
+                        description
+                        docs_url
+                        features {
+                            description
+                            key
+                            limit
+                            name
+                            note
+                            unit
+                        }
+                        free_allocation
+                        image_url
+                        included_if
+                        name
+                        plan_key
+                        product_key
+                        tiers {
+                            current_amount_usd
+                            current_usage
+                            flat_amount_usd
+                            unit_amount_usd
+                            up_to
+                        }
+                        unit
+                    }
+                }
+            }
+        }
+    }
+`
+
+export default function Plans({
+    groupsToShow,
+    showTitle,
+}: {
+    groupsToShow?: string[]
+    showTitle?: boolean
+}): JSX.Element {
+    const {
+        allProductData: {
+            nodes: [{ products }],
+        },
+    } = useStaticQuery(allProductsData)
+    return (groupsToShow?.length > 0 ? products.filter(({ type }) => groupsToShow.includes(type)) : products).map(
+        ({ type, plans, unit, addons, name, inclusion_only }) => {
+            return (
+                <div className="grid gap-y-4 min-w-[615px]" key={type}>
+                    {showTitle && <h4>{name}</h4>}
+                    {plans.some(({ free_allocation }) => free_allocation) && (
+                        <div>
+                            <Row className="bg-accent dark:bg-accent-dark mb-2">
+                                <Heading title="Plans" className="col-span-2" />
+                                {plans.map(({ free_allocation, plan_key }) => {
+                                    return (
+                                        <Heading
+                                            title={free_allocation ? 'Free' : 'Unlimited'}
+                                            subtitle={
+                                                free_allocation
+                                                    ? 'No credit card required'
+                                                    : 'All features, no limitations'
+                                            }
+                                            className="flex-shrink-0"
+                                            key={plan_key}
+                                        />
+                                    )
+                                })}
+                            </Row>
+                            <Row>
+                                <Title className="col-span-2" title={capitalize(`${unit}s`)} />
+                                {plans.map(({ free_allocation, plan_key }) => {
+                                    return (
+                                        <p key={plan_key} className="m-0 text-base opacity-70">
+                                            {free_allocation ? (
+                                                <>
+                                                    <strong>{free_allocation.toLocaleString()}</strong>
+                                                    <span className="text-xs">/mo</span>
+                                                </>
+                                            ) : (
+                                                <strong>Unlimited</strong>
+                                            )}
+                                        </p>
+                                    )
+                                })}
+                            </Row>
+                        </div>
+                    )}
+                    <div>
+                        <Row className="bg-accent dark:bg-accent-dark mb-2">
+                            <Heading title="Features" className="col-span-4" />
+                        </Row>
+                        {plans[plans.length - 1].features.map((feature, index) => {
+                            return (
+                                <Row className="hover:bg-accent/60 dark:hover:bg-accent-dark/70" key={feature.key}>
+                                    <div className="col-span-2">
+                                        <Tooltip
+                                            placement="right-end"
+                                            content={() => (
+                                                <div className="p-2 max-w-sm">
+                                                    <p className="font-bold text-[15px] mb-1">{feature.name}</p>
+                                                    <p className="mb-0 text-sm">{feature.description}</p>
+                                                </div>
+                                            )}
+                                        >
+                                            <span className="relative">
+                                                <Title
+                                                    className="border-b border-dashed border-border dark:border-dark inline-block cursor-default"
+                                                    title={feature.name}
+                                                />
+                                            </span>
+                                        </Tooltip>
+                                    </div>
+                                    {plans.map((plan) => (
+                                        <Feature key={`${feature.key}-${type}`} feature={plan.features?.[index]} />
+                                    ))}
+                                </Row>
+                            )
+                        })}
+                        {addons.map((addon) => {
+                            return (
+                                <Row className="hover:bg-accent/60 dark:hover:bg-accent-dark/70" key={addon.type}>
+                                    <div className="col-span-2">
+                                        <AddonTooltip addon={addon} parentProductName={name}>
+                                            <Title
+                                                className="border-b border-dashed border-border dark:border-dark inline-block cursor-default"
+                                                title={addon.name}
+                                            />
+                                            <Label className="ml-2" text="Addon" />
+                                        </AddonTooltip>
+                                    </div>
+                                    {plans.map((plan) => {
+                                        return plan.free_allocation ? (
+                                            <Close opacity={1} className="text-red w-4" />
+                                        ) : (
+                                            <AddonTooltip addon={addon} parentProductName={name}>
+                                                <Title
+                                                    className="border-b border-dashed border-border dark:border-dark inline-block cursor-default"
+                                                    title="Available"
+                                                />
+                                            </AddonTooltip>
+                                        )
+                                    })}
+                                </Row>
+                            )
+                        })}
+                    </div>
+                    <div>
+                        <Row className="bg-accent dark:bg-accent-dark mb-2">
+                            <Heading title="Monthly pricing" className="col-span-4" />
+                        </Row>
+                        <div>
+                            {inclusion_only ? (
+                                <InclusionOnlyRow plans={plans} />
+                            ) : (
+                                <PricingTiers plans={plans} unit={unit} />
+                            )}
+                        </div>
+                    </div>
+                    <Row>
+                        <div className="col-span-2" />
+                        {plans.map((_, i) => (
+                            <CTA key={`cta-${i}`} />
+                        ))}
+                    </Row>
+                </div>
+            )
+        }
+    )
+}
