+++ src/components/Product/ProductAnalytics/index.tsx
@@ -42,7 +42,7 @@ import Slider from 'components/Products/Slider'
 import MobileSlides from 'components/Products/MobileSlides'
 import { SEO } from 'components/seo'
 import { useLayoutData } from 'components/Layout/hooks'
-import PlanComparisonNew from 'components/Pricing/PlanComparisonNew'
+import Plans from 'components/Pricing/Plans'
 
 const product = {
     slug: 'product-analytics',
@@ -865,7 +865,7 @@ export const ProductProductAnalytics = () => {
 
                 <div className="md:flex justify-between items-start gap-12">
                     <div className="flex-grow overflow-auto">
-                        <PlanComparisonNew showHeaders={false} showCTA={false} groupsToShow={['product_analytics']} />
+                        <Plans showHeaders={false} showCTA={false} groupsToShow={['product_analytics']} />
                     </div>
                     <div className="md:w-96 md:mt-4">
                         <h4 className="text-3xl">FAQs</h4>
