+++ src/components/Product/Surveys/index.tsx
@@ -36,7 +36,7 @@ import { SmoothScroll } from 'components/Products/SmoothScroll'
 import { FAQ } from 'components/Products/FAQ'
 import { SEO } from 'components/seo'
 import { useLayoutData } from 'components/Layout/hooks'
-import PlanComparisonNew from 'components/Pricing/PlanComparisonNew'
+import Plans from 'components/Pricing/Plans'
 
 const product = {
     slug: 'surveys',
@@ -422,7 +422,7 @@ export const ProductSurveys = () => {
 
                 <div className="md:flex justify-between items-start gap-12">
                     <div className="flex-grow overflow-auto">
-                        <PlanComparisonNew showHeaders={false} showCTA={false} groupsToShow={['surveys']} />
+                        <Plans showHeaders={false} showCTA={false} groupsToShow={['surveys']} />
                     </div>
 
                     <div className="md:w-96 md:mt-4">
